クレジット決済機能のコンポーネント化
