+++ src/assets/css/input-check.css
@@ -1,0 +1,9 @@
+.ex {
+  font-size: 12px;
+  color: rgb(152, 152, 152);
+}
+
+.errorMessages {
+  font-size: 14px;
+  color: red;
+}
