+++ src/views/OrderConfirm.vue
@@ -6,6 +6,7 @@ import { useUserProviderKey } from "@/providers/useUserProvider";
 import axios from "axios";
 import { CartListKey } from "@/providers/useCartProvider";
 import OrderItemFormList from "@/components/OrderItemFormList.vue";
+import CreditCardPayment from "@/components/CreditCardPayment.vue";
 
 const name = ref("");
 const mailAddress = ref("");
@@ -28,18 +29,6 @@ const checkError = ref(true);
 const router = useRouter();
 const orderStore = inject(CartListKey);
 const userStore = inject(useUserProviderKey);
-
-const card_num = ref("");
-const card_exp_year = ref(0);
-const card_exp_month = ref(0);
-const card_name = ref("");
-const card_cvv = ref("");
-const errorMessageOfCreditCardNumber = ref("");
-const errorMessageOfCreditCardNumber2 = ref("");
-const errorMessageOfExpiry = ref("");
-const errorMessageOfCardName = ref("");
-const errorMessageOfCardCvv = ref("");
-const errorMessageOfNotNumber = ref("");
 
 if (!userStore) {
   throw new Error("");
@@ -161,113 +150,12 @@ const orderConfirm = async () => {
     checkError.value = true;
   }
 
-  // クレジットカードの決済処理
+  if (checkError.value === false) {
+    return;
+  }
+
   let currentUser = userStore.currentUser;
   let currentOrder = orderStore.userOrderInfo.value;
-
-  if (paymentMethod.value === "2") {
-    // エラーチェック
-    if (card_num.value === "") {
-      errorMessageOfCreditCardNumber.value =
-        "クレジットカード番号を入力してください";
-      checkError.value = false;
-    } else {
-      errorMessageOfCreditCardNumber.value = "";
-      checkError.value = true;
-    }
-
-    if (card_exp_year.value === 0 || card_exp_month.value === 0) {
-      errorMessageOfExpiry.value =
-        "クレジットカードの有効期限を入力してください";
-      checkError.value = false;
-    } else {
-      errorMessageOfExpiry.value = "";
-      checkError.value = true;
-    }
-
-    if (card_name.value === "") {
-      errorMessageOfCardName.value =
-        "クレジットカードの名義人を入力してください";
-      checkError.value = false;
-    } else {
-      errorMessageOfCardName.value = "";
-      checkError.value = true;
-    }
-
-    if (card_cvv.value === "") {
-      errorMessageOfCardCvv.value = "セキュリティコードを入力してください";
-      checkError.value = false;
-    } else {
-      errorMessageOfCardCvv.value = "";
-      checkError.value = true;
-    }
-
-    if (checkError.value === false) {
-      return;
-    }
-
-    const response2 = await axios.post(
-      "http://153.127.48.168:8080/sample-credit-card-web-api/credit-card/payment",
-      {
-        user_id: currentOrder.userId,
-        order_number: zipCode.value + paymentMethod.value + 12345678,
-        amount: currentOrder.calcTotalPrice,
-        card_number: card_num.value,
-        card_exp_year: card_exp_year.value,
-        card_exp_month: card_exp_month.value,
-        card_name: card_name.value,
-        card_cvv: card_cvv.value,
-      }
-    );
-    console.log(JSON.stringify(response2));
-
-    if (typeof card_num.value !== "number") {
-      errorMessageOfCreditCardNumber2.value =
-        "クレジットカード番号は数字で入力してください";
-      checkError.value = false;
-    } else {
-      errorMessageOfCreditCardNumber2.value = "";
-      checkError.value = true;
-    }
-
-    if (String(card_num.value).length !== 16) {
-      errorMessageOfCreditCardNumber.value =
-        "クレジットカード番号の桁数が間違っています";
-      checkError.value = false;
-    } else {
-      errorMessageOfCreditCardNumber.value = "";
-      checkError.value = true;
-    }
-
-    if (response2.data.error_code === "E-01") {
-      errorMessageOfExpiry.value = "有効期限がきれています";
-      checkError.value = false;
-    } else {
-      errorMessageOfExpiry.value = "";
-      checkError.value = true;
-    }
-
-    if (response2.data.error_code === "E-02") {
-      errorMessageOfCardCvv.value = "セキュリティーコードが間違っています";
-      checkError.value = false;
-    } else {
-      errorMessageOfCardCvv.value = "";
-      checkError.value = true;
-    }
-
-    if (response2.data.error_code === "E-03") {
-      errorMessageOfNotNumber.value =
-        "セキュリティーコードは数字で入力してください";
-      checkError.value = false;
-    } else {
-      errorMessageOfNotNumber.value = "";
-      checkError.value = true;
-    }
-
-    if (checkError.value === false) {
-      return;
-    }
-  }
 
   // 注文内容を送信する
   const response = await axios.post(
@@ -510,89 +398,7 @@ const getAddress = async () => {
 
       <div class="order-confirm-delivery-info">
         <div v-if="paymentMethod === '2'">
-          <div class="row">
-            <div class="errorMessages">
-              {{ errorMessageOfCreditCardNumber }}
-            </div>
-            <div class="errorMessages">
-              {{ errorMessageOfCreditCardNumber2 }}
-            </div>
-            <div class="input-field">
-              <label for="creditCardNumber">クレジットカード番号</label>
-              <input
-                type="text"
-                v-model.number="card_num"
-                id="creditCardNumber"
-                maxlength="16"
-              />
-              <div class="ex">例：1234 1234 5678 5678</div>
-            </div>
-          </div>
-          <div class="row">
-            <div class="errorMessages">
-              {{ errorMessageOfExpiry }}
-            </div>
-            <div class="col s10 expiry">
-              <span class="expiry2"> 有効期限： </span>
-              <select class="browser-default" v-model="card_exp_month">
-                <option value="1">1</option>
-                <option value="2">2</option>
-                <option value="3">3</option>
-                <option value="4">4</option>
-                <option value="5">5</option>
-                <option value="6">6</option>
-                <option value="7">7</option>
-                <option value="8">8</option>
-                <option value="9">9</option>
-                <option value="10">10</option>
-                <option value="11">11</option>
-                <option value="12">12</option></select
-              >月
-              <select class="browser-default" v-model="card_exp_year">
-                <option value="2022">2022</option>
-                <option value="2023">2023</option>
-                <option value="2024">2024</option>
-                <option value="2025">2025</option>
-                <option value="2026">2026</option>
-                <option value="2027">2027</option>
-                <option value="2028">2028</option>
-                <option value="2029">2029</option>
-                <option value="2030">2030</option>
-                <option value="2031">2031</option>
-                <option value="2032">2032</option>
-                <option value="2033">2033</option>
-                <option value="2034">2034</option>
-                <option value="2035">2035</option>
-                <option value="2036">2036</option>
-                <option value="2037">2037</option>
-                <option value="2038">2038</option></select
-              >年
-              <div class="ex">例：01月/2022年</div>
-            </div>
-          </div>
-          <div class="row">
-            <div class="errorMessages">
-              {{ errorMessageOfCardName }}
-            </div>
-            <div class="input-field">
-              <label for="card_name">カード名義人</label>
-              <input type="text" v-model="card_name" id="card_name" />
-            </div>
-            <div class="ex">例：TARO YAMADA</div>
-          </div>
-          <div class="row">
-            <div class="errorMessages">
-              {{ errorMessageOfCardCvv }}
-            </div>
-            <div class="errorMessages">
-              {{ errorMessageOfNotNumber }}
-            </div>
-            <div class="input-field">
-              <label for="card_cvv">セキュリティーコード</label>
-              <input type="password" v-model.number="card_cvv" id="card_cvv" />
-            </div>
-            <div class="ex">例：123</div>
-          </div>
+          <CreditCardPayment></CreditCardPayment>
         </div>
       </div>
 
@@ -608,13 +414,5 @@ const getAddress = async () => {
 </template>
 
 <style scoped>
-.ex {
-  font-size: 12px;
-  color: rgb(152, 152, 152);
-}
-
-.errorMessages {
-  font-size: 14px;
-  color: red;
-}
+@import url("@/assets/css/input-check.css");
 </style>
